Third argument to fcntl with F_SETFL should be an int anyhow.
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@2602 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/network/socket/socket_bsd.c
@@ -147,7 +147,7 @@ Socket_setNonBlocking(Socket *sock) {
 		return -1;
 	}
 
-	if (fcntl(sock->fd, F_SETFL, (long) (flags | O_NONBLOCK)) == -1) {
+	if (fcntl(sock->fd, F_SETFL, flags | O_NONBLOCK) == -1) {
 		int savedErrno = errno;
 		log_add(log_Error, "Setting non-blocking mode on socket failed: "
 				"%s.\n", strerror(errno));
